* linux build fixed
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4349 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-logonserver/AuthSocket.cpp
@@ -518,7 +518,7 @@ void AuthSocket::HandleReconnectChallenge()
 		return;
 
 	// Copy the data into our cached challenge structure
-	if((full_size+4) > sizeof(sAuthLogonChallenge_C))
+	if((size_t)(full_size+4) > sizeof(sAuthLogonChallenge_C))
 	{
 		Disconnect();
 		return;
